requirements.txt
Appium_Python_Client==4.1.0
Appium_Python_Client==4.4.0
Flask==3.1.0
selenium==4.27.1

Appium-Python-Client~=4.3.0
psutil~=5.9.5